+++ src/sentry/incidents/endpoints/serializers/workflow_engine_action.py
@@ -6,7 +6,9 @@ from sentry.incidents.endpoints.serializers.alert_rule_trigger_action import (
 )
 from sentry.incidents.models.alert_rule import AlertRuleTriggerAction
 from sentry.notifications.models.notificationaction import ActionService
-from sentry.workflow_engine.handlers.action.notification.base import MetricAlertRegistryInvoker
+from sentry.notifications.notification_action.group_type_notification_registry.handlers.metric_alert_registry_handler import (
+    MetricAlertRegistryHandler,
+)
 from sentry.workflow_engine.models import Action, ActionAlertRuleTriggerAction
 
 
@@ -20,7 +22,7 @@ class WorkflowEngineActionSerializer(Serializer):
         aarta = ActionAlertRuleTriggerAction.objects.get(action=obj.id)
         priority = obj.data.get("priority")
         type_value = ActionService.get_value(obj.type)
-        target = MetricAlertRegistryInvoker.target(obj)
+        target = MetricAlertRegistryHandler.target(obj)
 
         target_type = obj.config.get("target_type")
         target_identifier = obj.config.get("target_identifier")
