+++ src/sentry/notifications/notification_action/group_type_notification_registry/__init__.py
@@ -1,0 +1,7 @@
+__all__ = [
+    "IssueAlertRegistryHandler",
+    "MetricAlertRegistryHandler",
+]
+
+from .handlers.issue_alert_registry_handler import IssueAlertRegistryHandler
+from .handlers.metric_alert_registry_handler import MetricAlertRegistryHandler
