+++ src/sentry/notifications/notification_action/issue_alert_registry/__init__.py
@@ -1,7 +1,32 @@
-__all__ = ["issue_alert_handler_registry"]
+__all__ = [
+    "AzureDevopsIssueAlertHandler",
+    "DiscordIssueAlertHandler",
+    "EmailIssueAlertHandler",
+    "GithubIssueAlertHandler",
+    "GithubEnterpriseIssueAlertHandler",
+    "JiraIssueAlertHandler",
+    "JiraServerIssueAlertHandler",
+    "MSTeamsIssueAlertHandler",
+    "OpsgenieIssueAlertHandler",
+    "PagerdutyIssueAlertHandler",
+    "PluginIssueAlertHandler",
+    "SentryAppIssueAlertHandler",
+    "SlackIssueAlertHandler",
+    "WebhookIssueAlertHandler",
+    "PagerDutyIssueAlertHandler",
+]
 
-from sentry.utils.registry import Registry
-
-from .base import BaseIssueAlertHandler
-
-issue_alert_handler_registry = Registry[BaseIssueAlertHandler](enable_reverse_lookup=False)
+from .handlers.azure_devops_issue_alert_handler import AzureDevopsIssueAlertHandler
+from .handlers.discord_issue_alert_handler import DiscordIssueAlertHandler
+from .handlers.email_issue_alert_handler import EmailIssueAlertHandler
+from .handlers.github_enterprise_issue_alert_handler import GithubEnterpriseIssueAlertHandler
+from .handlers.github_issue_alert_handler import GithubIssueAlertHandler
+from .handlers.jira_issue_alert_handler import JiraIssueAlertHandler
+from .handlers.jira_server_issue_alert_handler import JiraServerIssueAlertHandler
+from .handlers.msteams_issue_alert_handler import MSTeamsIssueAlertHandler
+from .handlers.opsgenie_issue_alert_handler import OpsgenieIssueAlertHandler
+from .handlers.pagerduty_issue_alert_handler import PagerDutyIssueAlertHandler
+from .handlers.plugin_issue_alert_handler import PluginIssueAlertHandler
+from .handlers.sentry_app_issue_alert_handler import SentryAppIssueAlertHandler
+from .handlers.slack_issue_alert_handler import SlackIssueAlertHandler
+from .handlers.webhook_issue_alert_handler import WebhookIssueAlertHandler
