+++ src/sentry/notifications/notification_action/issue_alert_registry/handlers/jira_server_issue_alert_handler.py
@@ -1,9 +1,5 @@
-from sentry.notifications.notification_action.issue_alert_registry import (
-    issue_alert_handler_registry,
-)
-from sentry.notifications.notification_action.issue_alert_registry.base import (
-    TicketingIssueAlertHandler,
-)
+from sentry.notifications.notification_action.registry import issue_alert_handler_registry
+from sentry.notifications.notification_action.types import TicketingIssueAlertHandler
 from sentry.workflow_engine.models import Action
