+++ src/sentry/notifications/notification_action/issue_alert_registry/handlers/slack_issue_alert_handler.py
@@ -1,9 +1,7 @@
 from typing import Any
 
-from sentry.notifications.notification_action.issue_alert_registry import (
-    issue_alert_handler_registry,
-)
-from sentry.notifications.notification_action.issue_alert_registry.base import BaseIssueAlertHandler
+from sentry.notifications.notification_action.registry import issue_alert_handler_registry
+from sentry.notifications.notification_action.types import BaseIssueAlertHandler
 from sentry.workflow_engine.models import Action
 from sentry.workflow_engine.typings.notification_action import ActionFieldMapping, SlackDataBlob
