+++ src/sentry/notifications/notification_action/metric_alert_registry/__init__.py
@@ -1,7 +1,7 @@
-__all__ = ["metric_alert_handler_registry"]
+__all__ = [
+    "OpsgenieMetricAlertHandler",
+    "PagerDutyMetricAlertHandler",
+]
 
-from sentry.utils.registry import Registry
-
-from .base import BaseMetricAlertHandler
-
-metric_alert_handler_registry = Registry[BaseMetricAlertHandler](enable_reverse_lookup=False)
+from .handlers.opsgenie_metric_alert_handler import OpsgenieMetricAlertHandler
+from .handlers.pagerduty_metric_alert_handler import PagerDutyMetricAlertHandler
