+++ src/sentry/workflow_engine/handlers/action/notification/azure_devops_handler.py
@@ -1,6 +1,4 @@
-from sentry.workflow_engine.handlers.action.notification.ticketing.base import (
-    TicketingActionHandler,
-)
+from sentry.workflow_engine.handlers.action.notification.base import TicketingActionHandler
 from sentry.workflow_engine.models import Action
 from sentry.workflow_engine.registry import action_handler_registry
 from sentry.workflow_engine.types import ActionHandler
