+++ src/sentry/workflow_engine/handlers/action/notification/sentry_app_handler.py
@@ -1,9 +1,8 @@
 from sentry.notifications.notification_action.utils import execute_via_group_type_registry
-from sentry.workflow_engine.handlers.action.notification.base import ActionHandler
 from sentry.workflow_engine.handlers.action.notification.common import GENERIC_ACTION_CONFIG_SCHEMA
 from sentry.workflow_engine.models import Action, Detector
 from sentry.workflow_engine.registry import action_handler_registry
-from sentry.workflow_engine.types import WorkflowEventData
+from sentry.workflow_engine.types import ActionHandler, WorkflowEventData
 
 
 @action_handler_registry.register(Action.Type.SENTRY_APP)
